API /user/login

+++ Backend/routes/user/delete.py
@@ -8,9 +8,12 @@ from appInit import app, db, User, bcrypt
 
 def delete(data):
     userASuppr = User.query.filter_by(id=get_jwt_identity(data["token"])).first()
-
+    
+    if not userASuppr :
+        return jsonify({"status": "fail", "reason": "utilisateur inexistant"})
+    
     db.session.delete(userASuppr)
     db.session.commit()
 
     return jsonify({"status": "succes", "reason": "none"})
-    return jsonify({"status": "fail", "reason": "raison du fail"})
+
